change: move queue.py and song.py to cogs/ext
Following the move of db.py, queue.py and song.py are moved to ext/ as they are not cogs, and offer an interface for others to use.
